+++ CHANGELOG.md
@@ -14,6 +14,13 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [4.2.0] - 2022-08-23
+### Added
+- OAuthInteractive credential provider. This credential provider will redirect you to a login page
+and require that the user authenticates. It will also cache the token between runs.
+- OAuthDeviceCode credential provider. Display a device code to enter into a trusted device.
+It will also cache the token between runs.
+
 ## [4.1.2] - 2022-08-22
 ### Fixed
 - geospatial: support asset links for features